lib/bootstrap.sh: continue testing when fail to load rootfs partition
Most time rootfs partition is not a must for test, it's ok to move on.
One thing to do is it impacts test like phoronix, which is not
considered in this patch.

Two cases after this patch
1. switch from a general job to phoronix job (need rsync-rootfs) will
fail at the call of rsync_rootf, which impacts current job and phoronix
job
2. local boot switches to phoronix job, it fails at rsync_rootf and
reboot to get next job.

Both cases behaves differently without this patch, which is like
1. run a general job which fails at mount_rootfs, and reboot
to get next job (phoronix)
2. local boot switches to phoronix job and fails at mount_rootfs

+++ lib/bootstrap.sh
@@ -399,21 +399,27 @@ mount_rootfs()
 {
 	if [ -n "$rootfs_partition" ]; then
 		# wait for the machine to load the disk
-		wait_load_disk || {
-			# skipping following test if disk can't be load
-			echo "can't load the disk $rootfs_partition, skip testing..."
+		if wait_load_disk; then
+			ROOTFS_DIR=/opt/rootfs
+			mkdir -p $ROOTFS_DIR
+
+			mount $rootfs_partition $ROOTFS_DIR || {
+				mkfs.btrfs -f $rootfs_partition
+				mount $rootfs_partition $ROOTFS_DIR
+			}
+
+			mkdir -p $ROOTFS_DIR/tmp
+			CACHE_DIR=$ROOTFS_DIR/tmp
+			cleanup_pkg_cache $CACHE_DIR
+		else
+			# rli9 FIXME: continue the test unless the test requires to use rootfs
+			# on the rootfs_partition
+			echo "can't load the disk $rootfs_partition, try to continue the testing..."
 			set_job_state 'load_disk_fail'
-			return 1
-		}
-		ROOTFS_DIR=/opt/rootfs
-		mkdir -p $ROOTFS_DIR
-		mount $rootfs_partition $ROOTFS_DIR || {
-			mkfs.btrfs -f $rootfs_partition
-			mount $rootfs_partition $ROOTFS_DIR
-		}
-		mkdir -p $ROOTFS_DIR/tmp
-		CACHE_DIR=$ROOTFS_DIR/tmp
-		cleanup_pkg_cache $CACHE_DIR
+
+			CACHE_DIR=/tmp/cache
+			mkdir -p $CACHE_DIR
+		fi
 	else
 		CACHE_DIR=/tmp/cache
 		mkdir -p $CACHE_DIR
